services/knowledge-system/requirements.txt
# Knowledge System dependencies (kept in sync with actual imports)

# Core API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6

# HTTP / scraping
httpx>=0.25.0
trafilatura>=1.6.0
readability-lxml>=0.8.1
beautifulsoup4>=4.12.0
lxml>=4.9.0
DrissionPage>=4.0.0

# Config (.env support for shared.utils.config)
python-dotenv>=1.0.0

# AI client (used by shared.gemini)
google-genai>=1.0.0

# Monitoring / logging
prometheus-client>=0.18.0
python-json-logger>=2.0.7

# PDF extraction
pdfplumber>=0.10.0
pdf2image>=1.16.0

# OCR / image processing
rapidocr-onnxruntime>=1.3.0
pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0

# Deduplication
datasketch>=1.6.0
mmh3>=4.0.0

# File storage
minio>=7.2.0
